requirements.txt
aws-cdk-lib
constructs>=10.0.0
boto3
